% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{abiF}
\alias{abiF}
\title{abiF
Computes ABI for target F, e.g. ABImsy (Griffith et al. 2023)}
\usage{
abiF(stock, target = 0.2, threshold = 0.9, brp = NULL)
}
\arguments{
\item{stock}{object of class FLStock}

\item{target}{target F at equilibrium, e.g. Fmsy}

\item{threshold}{quantile ageref treshold, default 0.9}

\item{brp}{An optional input FLBRP object}
}
\value{
FLQuant
}
\description{
abiF
Computes ABI for target F, e.g. ABImsy (Griffith et al. 2023)
}
\examples{
data(ple4)
abimsy <- abiF(ple4, target=0.22, threshold=0.9)
plot(abimsy) + ylim(0, 2) +
 geom_hline(yintercept = 1) + ylab(expression(ABI[MSY]))
}
