% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{+,FLBRP,FLPar-method}
\alias{+,FLBRP,FLPar-method}
\title{Add external reference points to an FLBRP object}
\usage{
\S4method{+}{FLBRP,FLPar}(e1, e2)
}
\arguments{
\item{e1}{A \strong{FLBRP} object}

\item{e2}{A \strong{FLPar} containing a set of reference points.}
}
\value{
A \strong{FLBRP} object with the added and calculated reference points.
}
\description{
Reference points computed outside of \code{FLBRP} can be added to the \emph{@refpts}
slot of an \code{FLBRP} object. A subsequent call to \code{brp()} will compute all
other quantities related to this reference point.
}
\details{
The reference points to be added are passed as an \code{FLPar} object. Names of
these reference points need to follow this convention:
\itemize{
\item For SSB reference points, name should strat with "B", e.g. "Blim".
\item For F reference points, first letter must be "F", e.g. "Fmsy".
\item Yield reference points should start with "C", e.g. "Cpa".
\item For recruitment reference points, use "R" as starting letter, e.g. "R0".
}
}
\examples{
data(ple4brp)
refs <- FLPar(Fmsy=0.21, Blim=207288, Bpa=290203)
plot(ple4brp + refs)
}
\seealso{
\link{brp}
}
\author{
The FLR Team
}
\keyword{methods}
