% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/plots.R
\docType{methods}
\name{brp,FLBRP-method}
\alias{brp,FLBRP-method}
\alias{computeRefpts,FLBRP-method}
\alias{plot,FLBRP,missing-method}
\title{Fit an FLBRP object}
\usage{
\S4method{brp}{FLBRP}(object)

\S4method{computeRefpts}{FLBRP}(object)

\S4method{plot}{FLBRP,missing}(
  x,
  refpts = dimnames(x@refpts)$refpt,
  obs = FALSE,
  labels = TRUE,
  shapes = "missing",
  colours = "missing",
  panels = NULL,
  ncol = 2,
  removals = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{FLBRP}}

\item{x}{An object of class \code{FLBRP}}

\item{refpts}{Reference points to include, defaults to all in standard \code{refpts} slot bu \code{virgin}. A vector of class \code{character}}

\item{obs}{Should observations be plotted? Defaults to \code{FALSE}.}
}
\value{
An object of class \code{FLBRP} with reference points estimated

An object of class \link[=gg]{ggplot2}.
}
\description{
This method carries out the necessary calculation to obtain the reference
points for an object of class \code{FLBRP}. Results of the calculation are stored
in \verb{@refpts} slot of the object, and the object is then returned.

To directly obtain the recalculated \code{refpts} slot of an \code{FLBRP} object, the
\code{computeRefpts} method can be used. This is equivalent to \emph{fitting} the object
using \code{brp} and then extracting the \verb{@refpts} slot.

The plot method for the \code{FLBRP} class will show the relationship between
equilibrium SSB, Yield and Profit against levels of fishing mortality and
that of recruitment, Yield and profit against SSB.
}
\examples{
data(ple4brp)

ple4brp <- brp(ple4brp)
refpts(ple4brp)
m(ple4brp)[1:3,] <- 0.2
computeRefpts(ple4brp)
data(ple4brp)
plot(ple4brp)
plot(ple4brp, catch=TRUE)
# ADD observations
plot(ple4brp, obs=TRUE)
# SELECT which refpts to plot
plot(ple4brp, refpts=c("msy", "crash", "f0.1"))
plot(ple4brp, refpts=c("msy", "crash", "f0.1"), colours=c("green", "red", "blue"))
plot(ple4brp, refpts=c("msy", "crash", "f0.1"),
  colours=c("green", "red", "blue"), shapes=c(21, 22, 23))
# method will only plot existing referenced points
refpts(ple4brp) <- refpts(ple4brp)[c("msy", "fmax", "spr.30"),]
plot(ple4brp)
# Select which panels to plot
plot(ple4brp, panels=1:4)
# and in which format
plot(ple4brp, panels=1:3, ncol=1)
}
\seealso{
\link{FLBRP}

\link{FLBRP}
}
