\name{coerce}
\docType{methods}
\alias{coerce,FLlst,list-method}
\title{coercion of FLBRP objects}
\description{
  Methods to transform from and to class \code{\link{FLBRP-class}}.
}
\section{Generic function}{
 brp(object) 
}
\section{Methods}{\describe{
\item{}{}
  }
}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
