\name{computeRefpts}
\docType{methods}
\alias{computeRefpts}
\title{estimate BRPs}
\description{
  Estimates all specified reference points and returns an object of class \code{\link{refpts-class}}.
}
\section{Generic function}{
 computeRefpts(object) 
}
\section{Methods}{\describe{
\item{}{}
  }
}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
