\name{FLBRP-class}
\docType{class}
\alias{FLBRP-class}
\title{FLBRP class for biological reference points}
\description{
}
\section{Slots}{\describe{
\item{}{}
   }
}
\section{Extends}{\describe{
FLComp
}}
\section{Methods}{\describe{
\item{}{}}
}

\seealso{
}
\examples{
}
\references{
}
\author{}
\keyword{classes}
