\name{FLBRP-methods}
\docType{methods}
\alias{landings.n,FLBRP-method}
\title{methods for FLBRP}
\description{
  Various calculations based on the data available in an object of class \code{\link{FLBRP-class}}.
}
\section{Generic function}{
 brp(object) 
}
\section{Methods}{\describe{
\item{}{}
  }
}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
