% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/constructors.R
\docType{class}
\name{FLBRP}
\alias{FLBRP}
\alias{FLBRP-class}
\alias{FLBRP,missing,missing-method}
\alias{FLBRP-missing,missing-method}
\title{The FLR class for biological and economic reference points}
\usage{
\S4method{FLBRP}{missing,missing}(
  ...,
  model = formula(rec ~ a),
  params = FLPar(1, params = "a"),
  fbar = FLQuant(seq(0, 4, length.out = 101), quant = "age")
)
}
\description{
Estimation of equilibirum reference points on the results obtained from fitting
a population model can be carried out using the \code{FLBRP} class. Objects
of this classs contain the information necessary for the calculation of
population and fisheries quantities (abundance, catches, ...) under a range
of levels of fishing mortality ('F'). From these, the values of those
quantities that maximize or minimize the values of different indicators can
be calculated.
}
\details{
The most common input for the calculation of reference points is the result
of an stock assessment model fit, usually provided as two objects of class
\code{\link{FLStock}} and \code{\link{FLSR}}.
}
\section{Slots}{

\describe{
\item{\code{model}}{The formula for the stock-recruitment model, \verb{formula,}.}

\item{\code{params}}{Parameters of the stock-recruitment relationship, \verb{FLPar,}.}

\item{\code{refpts}}{The estimated reference points, \verb{FLPar,}.}

\item{\code{fbar}}{A vector of mean fishing mortality values for calculating the reference points, \verb{FLQuant,}.}

\item{\code{fbar.obs}}{The time series of observed mean fishing mortality, \verb{FLQuant,}.}

\item{\code{landings.obs}}{The time series of observed total landings, \verb{FLQuant,}.}

\item{\code{discards.obs}}{The time series of observed total discards, \verb{FLQuant,}.}

\item{\code{rec.obs}}{The time series of observed recruitment, \verb{FLQuant,}.}

\item{\code{ssb.obs}}{The time series of observed spawning stock biomasses, \verb{FLQuant,}.}

\item{\code{stock.obs}}{The time series of observed total stock abundance, \verb{FLQuant,}.}

\item{\code{profit.obs}}{The time series of profits, \verb{FLQuant,}.}

\item{\code{landings.sel}}{Selectivity at age for the landings, \verb{FLQuant,}.}

\item{\code{discards.sel}}{Selectivity at age for the discards, \verb{FLQuant,}.}

\item{\code{bycatch.harvest}}{X, Fishing mortality at age of bycatch species, \verb{FLQuant,}.}

\item{\code{stock.wt}}{Mean weight-at-age in the stock to use in calculations, \verb{FLQuant,}.}

\item{\code{landings.wt}}{Mean weight-at-age in the landings to use in calculations, \verb{FLQuant,}.}

\item{\code{discards.wt}}{Mean weight-at-age in the discards to use in calculations, \verb{FLQuant,}.}

\item{\code{bycatch.wt}}{Mean weight-at-age in the bycatch to use in calculations, \verb{FLQuant,}.}

\item{\code{m}}{Natural mortality at age, \verb{FLQuant,}.}

\item{\code{mat}}{Maturity at age, \verb{FLQuant,}.}

\item{\code{harvest.spwn}}{Proportion of harvest/fishing mortality before spawning, \verb{FLQuant,}.}

\item{\code{m.spwn}}{Proportion of natural mortality before spawning, \verb{FLQuant,}.}

\item{\code{availability}}{, \verb{FLQuant,}.}

\item{\code{price}}{Price at age, \verb{FLQuant,}.}

\item{\code{vcost}}{X, \verb{FLQuant,}.}

\item{\code{fcost}}{X, \verb{FLQuant,}.}
}}

\section{Validity}{


\describe{
\item{\code{.obs} slots dims 1:5}{Slots containing observations, named
\code{*.obs}, must share dimensions 1 to 5}
\item{by-age slots dims 1:5}{Slots containing vectors of values at age,
must share dimensions 1 to 5}
}

You can inspect the class validity function by using
\code{getValidity(getClassDef('FLBRP'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.

A standard method also exists for object of class \code{FLStock}, with or
without additional information on the stock-recruits relationship. This can
be provided as an object of class \code{FLSR}, one of class \code{predictModel}
or as a list wqith elements named 'params', of class \code{FLPar}, and 'model',
of class \code{formula}.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:

\describe{
\item{METHOD}{Neque porro quisquam est qui dolorem ipsum.}
}
}

\examples{

data(ple4)

# Create the FLSR and FLBRP objects
psr <- fmle(as.FLSR(ple4, model="bevholt"), control=list(trace=0))
prp <- brp(FLBRP(ple4, sr=psr))

summary(prp)
refpts(prp)
}
\seealso{
\link{FLComp} \link{FLStock} \link{FLSR} \link{brp}
}
\author{
The FLR Team
}
\keyword{classes}
