% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwdWindow.R
\docType{methods}
\name{fwdWindow,FLStock,FLBRP-method}
\alias{fwdWindow,FLStock,FLBRP-method}
\title{fwdWindow}
\usage{
\S4method{fwdWindow}{FLStock,FLBRP}(x, y, end = dims(x)$maxyear)
}
\arguments{
\item{x}{The population object, for example of class \code{\link{FLStock}}}

\item{y}{The reference points object, of class \code{\link{FLBRP}}}

\item{end}{Final year of the extended object, always interpreted as a \emph{character}}
}
\value{
An object of the same class a \emph{x}, extended to \emph{year=end}
}
\description{
Extends an object representing a fish population for projecting into the
future using the assumed equilibirum values used in the calculation of
reference points
}
\details{
\preformatted{slts <- c("stock.wt", "landings.wt", "discards.wt", "catch.wt",
   "m", "mat", "harvest.spwn", "m.spwn")
}
}
\examples{
data(ple4)

# Create the FLSR and FLBRP objects
psr <- fmle(as.FLSR(ple4, model="bevholt"), control=list(trace=0))
prp <- brp(FLBRP(ple4, sr=psr))

res <- fwdWindow(ple4, prp, end=2014)
}
