% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{msyRange,FLBRP-method}
\alias{msyRange,FLBRP-method}
\title{MSY range}
\usage{
\S4method{msyRange}{FLBRP}(object, Ftarget = "msy", range = 0.1)
}
\arguments{
\item{object}{\code{FLBRP}}

\item{range}{the values of yield around MSY}
}
\value{
object of class \code{FLPar} with reference point quantities
}
\description{
Calculates the values of fishing mortality, biomass, SSB etc. for a range of yields either side of MSY.
}
\examples{
data(ple4brp)

msyRange(ple4brp)
}
\seealso{
\code{\link{refpts}}
}
