% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-noise.R
\docType{methods}
\name{noise}
\alias{noise}
\alias{spectra}
\title{noise}
\usage{
noise(len, B = 0, trunc = 0, sd = 1)

spectra(x, fs = 1, norm = FALSE, pl = TRUE, omit = -(1:5))
}
\arguments{
\item{\code{n}, }{length of time series to generate}

\item{\code{B}, }{autocorrelation, a real number in [0,1]}

\item{\code{trunc}, }{get rid of first values equal to trunc, i.e. to allow burn in}

\item{\code{sd}, }{standard deviation}

\item{\code{n}, }{length of time series to generate}

\item{\code{B}, }{autocorrelation, a real number in [0,1]}

\item{\code{trunc}, }{get rid of first values equal to trunc, i.e. to allow burn in}

\item{\code{sd}, }{standard deviation}
}
\value{
A vector with autocorrelation equal to B and of length \code{n-trunc}

A vector with autocorrelation equal to B and of length \code{n-trunc}
}
\description{
A noise generator

A noise generator
}
\examples{
white <- noise(1000,B=0)
plot(white)
acf(white)

red <- noise(1000,B=0.8)
plot(red)
acf(white)

blue <- noise(1000,B=-0.8)
plot(blue)
acf(blue)

white <- noise(1000,B=0)
plot(white)
acf(white)

red <- noise(1000,B=0.8)
plot(red)
acf(white)

blue <- noise(1000,B=-0.8)
plot(blue)
acf(blue)

}
\references{
Ranta and Kaitala 2001 Proc. R. Soc.
vt = B * vt-1 + s * sqrt(1 - B^2)
s is normally distributed random variable with mean = 0, variance = 1
B is the autocorrelation parameter

Ranta and Kaitala 2001 Proc. R. Soc.
vt = B * vt-1 + s * sqrt(1 - B^2)
s is normally distributed random variable with mean = 0, variance = 1
B is the autocorrelation parameter
}
