% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{r,FLBRP,missing-method}
\alias{r,FLBRP,missing-method}
\title{Intrisic rate of increase}
\usage{
\S4method{r}{FLBRP,missing}(m, by = "year", method = "el", ...)
}
\arguments{
\item{m}{An object of class \code{FLBRP}}
}
\value{
object of class \code{FLQuant} with \emph{r} estimate
}
\description{
Calculates the value of \emph{r}, the intrisic rate of increase.
}
\examples{
data(ple4brp)

r(ple4brp)
}
\seealso{
\code{\link[FLCore:r]{FLCore::r()}}
}
