% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remap}
\alias{remap}
\title{A function to reshape the refpts slot of an FLBRP object.}
\usage{
remap(
  object,
  map = list(FMSY = c("msy", "harvest"), SBMSY = c("msy", "ssb"), BMSY = c("msy",
    "biomass"), B0 = c("virgin", "biomass"), SB0 = c("virgin", "ssb")),
  ...
)
}
\arguments{
\item{object}{An FLPar with dimensions 'refpt', 'quant' and 'iter'.}

\item{map}{A names list of dimnames to extract each reference point from.}

\item{...}{Extra reference points to be extracted.}
}
\value{
An FLPar object with dimensions 'params' and 'iter'.
}
\description{
The FLPar containing all estimated reference points in the 'refpts' slot
of an object of class 'FLBRP' can be converted into a simpler format with
a selection of relevant points. This makes it easier to use the object for
plotting and computation. Individual reference points, for example, can be
accessed using the '$' operator. The examples below show the difference
between both formats.
}
\details{
A series of basic reference poiunts are extracted by default with the following names:
\itemize{
\item FMSY: fishing mortality at MSY, c("msy", "harvest").
\item SBMSY: spawning biomass at MSY, c("msy", "ssb").
\item BMSY: total biomass at MSY, c("msy", "biomass").
\item B0: virgin biomass, c("virgin", "biomass").
\item SB0: virgin spawning biomass, c("virgin", "ssb").
}
}
\examples{
data(ple4brp)
# Input is refpts slot in FLBRP
refpts(ple4brp)
# Call with default map
remap(refpts(ple4brp))
# Additional refpts can be extracted
remap(refpts(ple4brp), FMEY=c("mey", "harvest"))
}
\seealso{
\link{FLBRP} \link{refpts}
}
\author{
Iago Mosqueira (WMR)
}
\keyword{classes}
