% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{segRef}
\alias{segRef}
\title{Reference points associated with inflection point of a segmented regression}
\usage{
segRef(x, ssb = params(x)["b"])
}
\arguments{
\item{x}{An object of class FLBRP.}

\item{ssb}{The SSB value of the inflection point, defaults to the 'b' parameter of the 'params' slot.}
}
\value{
An object of class 'FLBRP' with a new 'segreg' row in 'refpts'.
}
\description{
Segmented regression (a.k.a. hockey-stick) stock-recruitment relationships
are often used in short-term forecasts and when other SRRs can not be
applied. This function assigns the 'b' parameter of the relationship as an
SSB-related reference point and then calculates all other related ones at e
equilibrium. The functions returns an FLBRP object with the new refpts,
called 'segreg'.
The object must be set with a fitted 'segreg' stock-recruit relationship.
}
\examples{
data(ple4brp)
sr(ple4brp) <- sr(ple4brp, model='segreg')
refpts(segRef(ple4brp))
}
\seealso{
\linkS4class{FLBRP} \code{\link[FLCore:SRModels]{FLCore::segreg()}}
}
\author{
The FLR Team
}
\keyword{classes}
