% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sp,FLBRP,missing,ANY-method}
\alias{sp,FLBRP,missing,ANY-method}
\title{Surplus production}
\usage{
\S4method{sp}{FLBRP,missing,ANY}(stock, rel = TRUE)
}
\arguments{
\item{m}{An object of class \code{FLBRP}}
}
\value{
object of class \code{FLQuant} with \emph{sp} estimate
}
\description{
Calculates the surplus production.
}
\examples{
data(ple4brp)

sp(ple4brp)
}
\seealso{
\code{\link[FLCore:sp]{FLCore::sp()}}
}
