% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spr.R
\docType{methods}
\name{spr,FLBRP-method}
\alias{spr,FLBRP-method}
\alias{spr0,FLBRP,missing,missing-method}
\title{Spawners per recruit}
\usage{
\S4method{spr}{FLBRP}(object)

\S4method{spr0}{FLBRP,missing,missing}(ssb)
}
\arguments{
\item{object}{An object of class \code{FLBRP}}
}
\value{
An object of class \code{FLQuant}.
}
\description{
Calculates the spawners per recruit (\code{spr}) at each level of fishing mortality (F)
in the \code{fbar} slot, or the spawners per recruit when F=0 (\code{spr0}).
}
\examples{
data(ple4brp)

spr(ple4brp)
spr0(ple4brp)
data(ple4brp)

spr0(ple4brp)
}
