% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs-refpts.R
\docType{methods}
\name{msyRange,FLBRP-method}
\alias{msyRange,FLBRP-method}
\title{MSY range}
\usage{
\S4method{msyRange}{FLBRP}(object, range = 0.1)
}
\arguments{
\item{object}{\code{FLBRP}}

\item{range}{the values of yield around MSY}
}
\value{
object of class \code{FLPar} with reference point quantities
}
\description{
Calculates the values of fishing mortality, biomass, SSB etc. for a range of yields either side of MSY.
}
\examples{
\dontrun{
library(FLBRP)

data(ple4)
ple4.sr=fmle(as.FLSR(ple4,model="bevholt"),control=list(silent=TRUE))

ple4.eql=FLBRP(ple4,sr=ple4.sr)

msyRange(ple4.eql)
}
}
\seealso{
\code{\link{refpts}}
}

