\name{refpts-class}
\docType{class}
\alias{refpts-class}
\title{refpts class for reference points}
\description{
}
\section{Slots}{\describe{
\item{}{}
  }
}
\section{Extends}{\describe{
FLPar
array
}}
\section{Methods}{\describe{
\item{}{}
  }
}

\seealso{
}
\examples{
}
\references{
}
\author{}
\keyword{classes}
